better kmalloc and kfree !

+++ src/mm/malloc.h
@@ -1,12 +1,9 @@
-/* MALLOC.C */
-/* Copyright (c) @asdasda3456 2024 - 2024 */
-
 #ifndef MALLOC_H
 #define MALLOC_H
 
 void kmallocInit(uint32_t heapSize);
 void changeHeapSize(int newSize);
-void* kmalloc(uint32_t size);
+void* kmalloc(size_t size);
 void kfree(void* ptr);
 
 #endif
